Added $postId as parameter to the vhp_purge_urls filter
Just something quick before this is released.  This is helpful in order to differentiate what extra URLs to add. Say, the $postId is a product, then purge the shop home too and all upselling products. If it is a blog, do not purge anything related to shop, just the blog stuff. Depending on a users need there might be even more. So passing the $postId really helps.