add do_actoin after full cache purge
To allow other plugins to act on a full cache purge (e.g. Autoptimize so it can purge it's own cache). it might be preferable to call the action differently (e.g. `varnishttp_after_full_purge` to avoid conflicts with actions from other plugins (one never knows)) but I stuck to the naming in the other action.